+++ src/data/packer_template.go
@@ -83,9 +83,6 @@ func (j *JsonDatabase) AddPackerTemplate(ctx basecontext.ApiContext, template *m
 
 	j.data.PackerTemplates = append(j.data.PackerTemplates, *template)
 
-	if err := j.Save(ctx); err != nil {
-		return nil, err
-	}
 	return template, nil
 }
 
@@ -105,9 +102,7 @@ func (j *JsonDatabase) DeletePackerTemplate(ctx basecontext.ApiContext, nameOrId
 			}
 
 			j.data.PackerTemplates = append(j.data.PackerTemplates[:i], j.data.PackerTemplates[i+1:]...)
-			if err := j.Save(ctx); err != nil {
-				return err
-			}
+
 			return nil
 		}
 	}
@@ -126,9 +121,7 @@ func (j *JsonDatabase) UpdatePackerTemplate(ctx basecontext.ApiContext, template
 				return nil, ErrUpdatingInternalPackerTemplate
 			}
 			j.data.PackerTemplates[i] = *template
-			if err := j.Save(ctx); err != nil {
-				return nil, err
-			}
+
 			return template, nil
 		}
 	}