+++ src/orchestrator/common.go
@@ -11,6 +11,8 @@ const (
 
 func (s *OrchestratorService) getApiClient(request models.OrchestratorHost) *apiclient.HttpClientService {
 	apiClient := apiclient.NewHttpClient(s.ctx)
+	apiClient.WithHeader("X-SOURCE", "ORCHESTRATOR_REQUEST")
+
 	if request.Authentication != nil {
 		if request.Authentication.ApiKey != "" {
 			apiClient.AuthorizeWithApiKey(request.Authentication.ApiKey)